Created a list of categories; started work on implementing feature where users can see how much their total budget is after they allocate money towards a category
